organize blocks into folders
