+++ app/blocks/logic/compare.js
