+++ app/blocks/logic/negate.js
