+++ app/blocks/logic/ternary.js
