+++ app/blocks/text/compare.js
