+++ app/extensions/example_extension.js
